+++ packages/hugin/canvas/hugin_canvas.ml
@@ -1,0 +1,9 @@
+(*---------------------------------------------------------------------------
+  Copyright (c) 2026 The Raven authors. All rights reserved.
+  SPDX-License-Identifier: ISC
+  ---------------------------------------------------------------------------*)
+
+module Color = Color
+module Affine = Affine
+module Path = Path
+module Stroke = Stroke